fixed warnings in map2cs
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13552 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/import/map2cs/texfile.cpp
@@ -108,7 +108,7 @@ bool CTextureFile::AddToVFS(csRef<iVFS> VFS, const char* path)
   {
     size_t vfssize;
     VFS->GetFileSize (m_Filename, vfssize);
-    if (vfssize == m_OriginalData.GetSize())
+    if (vfssize == (size_t) m_OriginalData.GetSize())
     {
       res = true;
     }
